+++ src/spatialdata/__init__.py
@@ -11,12 +11,14 @@ __all__ = [
     "dataloader",
     "concatenate",
     "rasterize",
+    "rasterize_bins",
     "to_circles",
     "to_polygons",
     "transform",
     "aggregate",
     "bounding_box_query",
     "polygon_query",
+    "get_element_instances",
     "get_values",
     "join_spatialelement_table",
     "match_element_to_table",
@@ -39,10 +41,12 @@ from spatialdata._core.concatenate import concatenate
 from spatialdata._core.data_extent import are_extents_equal, get_extent
 from spatialdata._core.operations.aggregate import aggregate
 from spatialdata._core.operations.rasterize import rasterize
+from spatialdata._core.operations.rasterize_bins import rasterize_bins
 from spatialdata._core.operations.transform import transform
 from spatialdata._core.operations.vectorize import to_circles, to_polygons
 from spatialdata._core.query._utils import get_bounding_box_corners
 from spatialdata._core.query.relational_query import (
+    get_element_instances,
     get_values,
     join_spatialelement_table,
     match_element_to_table,