+++ tests/core/query/test_relational_query.py
@@ -520,7 +520,7 @@ def test_locate_value(sdata_query_aggregation):
     )
 
 
-def test_get_values_df(sdata_query_aggregation):
+def test_get_values_df_shapes(sdata_query_aggregation):
     # test with a single value, in the dataframe; using sdata + element_name
     v = get_values(
         value_key="numerical_in_gdf", sdata=sdata_query_aggregation, element_name="values_circles", table_name="table"
@@ -621,10 +621,55 @@ def test_get_values_df(sdata_query_aggregation):
         )
 
 
+def test_get_values_df_points(points):
+    # testing get_values() for points, we keep the test more minimalistic than the one for shapes
+    p = points["points_0"]
+    p = p.drop("instance_id", axis=1)
+    p.index.compute()
+    n = len(p)
+    obs = pd.DataFrame(index=p.index, data={"region": ["points_0"] * n, "instance_id": range(n)})
+    obs["region"] = obs["region"].astype("category")
+    table = TableModel.parse(
+        AnnData(shape=(n, 0), obs=obs), region="points_0", region_key="region", instance_key="instance_id"
+    )
+    points["points_0"] = p
+    points["table"] = table
+
+    assert get_values(value_key="region", element_name="points_0", sdata=points, table_name="table").shape == (300, 1)
+    get_values(value_key="instance_id", element_name="points_0", sdata=points, table_name="table")
+    get_values(value_key=["x", "y"], element_name="points_0", sdata=points, table_name="table")
+    get_values(value_key="genes", element_name="points_0", sdata=points, table_name="table")
+
+    pass
+
+
+def test_get_values_obsm(adata_labels: AnnData):
+    get_values(value_key="tensor", element=adata_labels)
+
+    get_values(value_key=["tensor", "tensor_copy"], element=adata_labels)
+
+    values = get_values(value_key="tensor", element=adata_labels, return_obsm_as_is=True)
+    assert isinstance(values, np.ndarray)
+
+
+def test_get_values_table(sdata_blobs):
+    df = get_values(value_key="channel_0_sum", element=sdata_blobs["table"])
+    assert isinstance(df, pd.DataFrame)
+    assert len(df) == 26
+
+
+def test_get_values_table_element_name(sdata_blobs):
+    sdata_blobs["table"].obs["region"] = sdata_blobs["table"].obs["region"].cat.add_categories("another_region")
+    sdata_blobs["table"].obs.loc["1", "region"] = "another_region"
+    sdata_blobs["table"].uns["spatialdata_attrs"]["region"] = ["blobs_labels", "another_region"]
+    sdata_blobs["another_region"] = sdata_blobs["blobs_labels"]
+    df = get_values(value_key="channel_0_sum", element=sdata_blobs["table"], element_name="blobs_labels")
+    assert isinstance(df, pd.DataFrame)
+    assert len(df) == 25
+
+
 def test_get_values_labels_bug(sdata_blobs):
     # https://github.com/scverse/spatialdata-plot/issues/165
-    from spatialdata import get_values
-
     get_values("channel_0_sum", sdata=sdata_blobs, element_name="blobs_labels", table_name="table")